add node

+++ 0x1A-hash_tables/3-hash_table_set.c
@@ -14,7 +14,7 @@ int hash_table_set(hash_table_t *ht, const char *key, const char *value)
 {
 	hash_node_t *hn;
 	hash_node_t *tmp;
-	int i = 0;
+	unsigned long int i = 0;
 
 	if (ht == NULL || key == NULL)
 		return (0);
